// src/features/blog/models/post.dto.ts
import { Audit } from '@/common/models/auditing.domain';
import { UserDto } from '@/features/user/models/user.dto';
import { PostStatus } from './post-status.enum';
import { PostAccess } from './post-access.enum';
import { TagDto } from './tag.dto';

export class PostDto {
  id: number;
  cover?: string;
  title: string;
  slug: string;
  excerpt: string;
  body?: string;
  status: PostStatus;
  access: PostAccess;
  featured: boolean;
  publishedAt: number;
  author: UserDto;
  tags: TagDto[];
  audit?: Audit;

  constructor(partial: Partial<PostDto> = {}) {
    Object.assign(this, partial);
  }
}